services/notifier/requirements.txt
fastapi
pydantic
requests
boto3 
uvicorn[standard]
python-multipart
python-dotenv
